fix: Don't ignore Yarn patch files (#25398)
An earlier commit modified `.gitignore` to ignore `.yarn` files in
`development/generate-attributions` (which is a special sub-project in
this repo used in the attribution generation workflow). However, this
change ended up ignoring Yarn patch files in `.yarn/patches` (along with
possibly other Yarn-specific files). This commit narrows the pattern in
`.gitignore` not to do this.