fix: flaky test `Settings Redirects to ENS domains when user inputs ENS into address bar` (#24898)
This PR fixes some flakiness of the test `Settings Redirects to ENS
domains when user inputs ENS into address bar`. After looking into the
code I've identified a source of flakiness, as well as a problem that
prevents to upload the artifacts (screenshot) if the test fail.

- The test was not using any mock: this is a source of flakiness, and
it's now fixed by adding the necessary mocks to the ENS contracts and
Infura calls (we need to set Mainnet, since this feature is only
available on Mainnet)
- The test was not using fixtures: this per se is not a problem, but the
issue here is that we are not triggering the
`verboseReportOnFailure(title,error)` on error, making it harder to
debug, as no artifacts are uploaded for this test, if it fails. Now we
are going to be able to see the screenshot if the test ever fails again

Circle ci job:
https://app.circleci.com/pipelines/github/MetaMask/metamask-extension/83569/workflows/0883aa51-cd94-4066-9f96-77d1375e5e9b/jobs/2994904/tests

Failure log: `TimeoutError: Wait timed out after 213ms`

Extra note: @davidmurdoch is also working on mitigating the rest of the
flakiness by waiting for the metamask to be fully initialized before
proceeding with the test. This work will be done in a separate PR.

---------

Co-authored-by: Dan J Miller <danjm.com@gmail.com>
Co-authored-by: Harika Jetpoluru <153644847+hjetpoluru@users.noreply.github.com>
Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>
Co-authored-by: Howard Braham <howrad@gmail.com>