Patch selected network controller for v12.0.0 (#25554)
This patch reapplies the changes previously patched on v11.16.8:
https://github.com/MetaMask/metamask-extension/pull/25127

This patch needed to be updated for v13 of the selected network
controller. I made a mistake in merge conflict resolution and am instead
updating these packages in this separate commit.

The core branch that was used to generate the patch can be seen here:
https://github.com/MetaMask/core/pull/new/patch-selected-network-controller-13.0.0-setNetworkClient-guard

~~This patch reapplies the changes previously patched on v11.16.7 and
v11.16.8:
https://github.com/MetaMask/metamask-extension/pull/25046
https://github.com/MetaMask/metamask-extension/pull/25127

These patches needed to be updated for v13 of the selected network
controller. I made a mistake in merge conflict resolution and am instead
updating these packages in this separate commit.

The core branch that was used to generate the patch can be seen here:
https://github.com/MetaMask/core/pull/new/patch-selected-network-controller-13.0.0-setNetworkClient-guard~~